+++ src/core/GameObject.cpp
@@ -1,0 +1,14 @@
+/**
+ * @file Entity.cpp
+ * @brief Implementation file for the Entity class.
+ *
+ * This file contains the implementation of the Entity class which is responsible for managing the entities in the game.
+ * The Entity class is responsible for adding, removing, and updating the components of the entities in the game.
+ *
+ * @author Nur Akmal bin Jalil
+ * @date 2024-08-12
+ */
+
+#include "GameObject.h"
+
+